implement colocation for team and cleaning up and renaming id to teamId in teamIdSchema
