+++ src/pages/addLeagueMatch/index.tsx
@@ -9,7 +9,7 @@ export default function AddMatch() {
   const { teamId } = useContext(TeamContext);
   const { leagueId } = useContext(LeagueContext);
   const { data, isLoading } = api.league.find.useQuery({
-    id: teamId,
+    teamId: teamId,
     leagueId,
   });
