--- src/server/types/officeTypes.ts
@@ -1,8 +1,0 @@
-export const offices = {
-  All: "All",
-  Oslo: "Oslo",
-  Trondheim: "Trondheim",
-  Bergen: "Bergen",
-} as const;
-
-export type Office = keyof typeof offices | "";
